mle/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
anyio==3.6.2
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constants==0.6.0
credentials==1.1
cryptography==41.0.1
fastapi==0.95.2
Flask==2.3.2
h11==0.14.0
httpcore==0.17.1
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.2
openpyxl==3.1.2
pandas
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
requests==2.30.0
services==0.1.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0
Werkzeug==2.3.4
# Query insights packages
git+https://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@github.com/tigerrepository/Insights-Pro.git@dev#egg=query_insights
# openai==0.27.2
# tiktoken==0.3.3
# gensim==4.2.0
# numpy==1.22.0
# transformers==4.22.0
# nltk==3.7
# kaleido==0.0.1
# mysql-connector-python==8.0.32
# statsmodels==0.13.5
# plotly==5.9.0
# gradio==3.25.0
# scikit-learn==1.2.2
# seaborn==0.12.2
# pandas==1.4.2
# streamlit==1.21.0
# nbformat==5.9.0
# adlfs==2023.4.0